Transportation Simulation
Highly realistic objects and environments can be generated by 3D
computer graphics and visualization. This technology is normally used in
various applications such as video games and virtual environments for
simulations. The major objective of this program is to develop a
transportation simulation, in which consists of a simplified
intersection (including traffic lights and surveillance cameras) and a
controllable vehicle. In this program, programmers need to develop an
entire control system of traffic signals, car driving simulation, and
multiple camera views of the simulation environment through different
perspectives. Moreover, in order to generate a realistic simulation
environment, a large amount of low-poly 3D models are added to this
program. With a great amount of low-poly 3D models, a realistic
simulation environment could be generated.